do not allow app bar to animate on navigating
